tidying up

+++ bin/macos/install.zsh
@@ -9,39 +9,55 @@
 ##   ##:::: ##: ##:::: ##:. ######::. #######::. ######::
 ##  ..:::::..::..:::::..:::......::::.......::::......:::
 
-echo "\n<<< Starting macOS Setup >>>\n"
+while true; do
 
-osascript -e 'tell application "System Preferences" to quit'
+    echo "Preparing to setup macOS, would you like to proceed (y/n)?"
 
-# Finder > View > Show Path Bar
-defaults write com.apple.finder ShowPathbar -bool true
+    read choice
 
-# Finder > Preferences > General > New Finder windows show:
-defaults write com.apple.finder NewWindowTarget -string 'PfLo'
-defaults write com.apple.finder NewWindowTargetPath -string "file://$HOME/.dotfiles"
+    case $choice in
+        y)
+            echo "Starting macOS Setup."
 
-# System Preferences > Dock
-defaults write com.apple.dock magnification -bool true
-defaults write com.apple.dock tilesize -int 45
-defaults write com.apple.dock largesize -int 60
-defaults write com.apple.dock autohide -bool true
-defaults write com.apple.dock autohide-time-modifier -float 0.25
-defaults write com.apple.dock autohide-delay -float 0.1
+            osascript -e 'tell application "System Preferences" to quit'
 
-# System Preferences > Accessibility > Pointer Control > Mouse & Trackpad > Trackpad Options > Enable Dragging > Three Finger Drag (NOTE: The GUI doesn't update)
-defaults write com.apple.AppleMultitouchTrackpad TrackpadThreeFingerDrag -bool true
+            # Finder > View > Show Path Bar
+            defaults write com.apple.finder ShowPathbar -bool true
 
+            # Finder > Preferences > General > New Finder windows show:
+            defaults write com.apple.finder NewWindowTarget -string 'PfLo'
+            defaults write com.apple.finder NewWindowTargetPath -string "file://$HOME/.dotfiles"
 
-# Third-Party Software
+            # System Preferences > Dock
+            defaults write com.apple.dock magnification -bool true
+            defaults write com.apple.dock tilesize -int 45
+            defaults write com.apple.dock largesize -int 60
+            defaults write com.apple.dock autohide -bool true
+            defaults write com.apple.dock autohide-time-modifier -float 0.25
+            defaults write com.apple.dock autohide-delay -float 0.1
 
-# iTerm2 Settings
-defaults write com.googlecode.iterm2 LoadPrefsFromCustomFolder -bool true
-defaults write com.googlecode.iterm2 PrefsCustomFolder -string "$HOME/.dotfiles/iterm2"
-defaults write com.googlecode.iterm2 NoSyncNeverRemindPrefsChangesLostForFile -bool true
+            # System Preferences > Accessibility > Pointer Control > Mouse & Trackpad > Trackpad Options > Enable Dragging > Three Finger Drag (NOTE: The GUI doesn't update)
+            defaults write com.apple.AppleMultitouchTrackpad TrackpadThreeFingerDrag -bool true
 
+            # Third-Party Software
 
-# Finish macOS Setup
-killall Finder
-killall Dock
-echo "\n<<< macOS Setup Complete.
-    A logout or restart might be necessary. >>>\n"
+            # iTerm2 Settings
+            defaults write com.googlecode.iterm2 LoadPrefsFromCustomFolder -bool true
+            defaults write com.googlecode.iterm2 PrefsCustomFolder -string "$HOME/.dotfiles/iterm2"
+            defaults write com.googlecode.iterm2 NoSyncNeverRemindPrefsChangesLostForFile -bool true
+
+            # Finish macOS Setup
+            killall Finder
+            killall Dock
+            echo "macOS Setup Complete, A logout or restart might be necessary."
+            break
+            ;;
+        n)
+            echo "Skipping macOS Setup."
+            exit 0
+            ;;
+        *)
+            echo "Invalid input. Please enter 'y' or 'n'."
+            ;;
+    esac
+done
